Replace auto_login file with inline code
